requirements.txt
Flask==0.12.2
requests==2.11.1
mysql-connector==2.1.4